requirements.txt
# ── RF capture ─────────────────────────────────────────────────────────────
scapy>=2.5.0

# ── Web framework ──────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
aiofiles>=23.2.0
python-multipart>=0.0.9

# ── GPS ────────────────────────────────────────────────────────────────────
pyserial>=3.5
pyserial-asyncio>=0.6
pynmea2>=1.19.0

# ── Config / CLI ───────────────────────────────────────────────────────────
pyyaml>=6.0.1
click>=8.1.7

# ── Terminal UI ────────────────────────────────────────────────────────────
rich>=13.7.0

# ── Misc ───────────────────────────────────────────────────────────────────
python-dateutil>=2.9.0
